Actually dump the backup GPT, not a copy of the primary GPT partitions
The backup GPT is a copy of the primary GPT header and partition
entries in reverse order at the end of the disk.

At the moment "edl.py gpt" dumps a copy of the primary GPT partition
to "gpt_backup0.bin", so it is missing a copy of the backup GPT header.

Change the code to actually make use of the fh.get_backup_gpt() function.

Also fix the starting sector in get_backup_gpt(): header["backup_lba"]
is the sector of the GPT backup header, which comes after the copy
of the partition entries. Usually the GPT backup header should be
the last 33 sectors after last_usable_lba, so use that instead.