Add SDX75 keys
Netgear has released a family of new M7 Pro devices (such as the MR7400-1A1NAS). I add new generator keys extracted from the firmware MR7400-1A1NAS_23115898_NTGX75_10.03.22.01_00_ATT_02.11.
The key for !OPENLOCK and !OPENCND is tested on a real device. The key for !OPENMEP could not be verified because the command is blocked on AT&T devices